Move dexcom driver, pt 2
