+++ lib/core/device.js
@@ -26,7 +26,7 @@ var driverManager = require('../driverManager');
 var builder = require('../objectBuilder')();
 var timeutils = require('../timeutils');
 
-var dexcomDriver = require('../dexcomDriver');
+var dexcomDriver = require('../drivers/dexcomDriver');
 var oneTouchMiniDriver = require('../oneTouchMiniDriver');
 var abbottFreeStyle = require('../abbottFreeStyle');
