+++ lib/drivers/dexcomDriver.js
@@ -20,10 +20,10 @@ var async = require('async');
 var moment = require('moment-timezone');
 var sundial = require('sundial');
 
-var crcCalculator = require('./crc.js');
-var struct = require('./struct.js')();
+var crcCalculator = require('../crc.js');
+var struct = require('../struct.js')();
 
-var debug = require('./bows')('DexcomDriver');
+var debug = require('../bows')('DexcomDriver');
 
 module.exports = function (config) {
   var cfg = _.clone(config);
